// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/shared/models/bankReceiptInvoiceMappingCustomerOrder.model.ts
export class PhieuThuBaoCoMappingCustomerOrderModel {
    id: string;
    orderCode: string;
    objectId: string | null;
    objectType: number | null;
    orderId: string | null;
    vendorOrderId: string | null;
    amount: number | null;
    listPacketServiceName: string;
    orderTypeName: string;
    createdDate: string | null;
}